All code/logic references to T1/T2/T2.5/T3 have been fully refactored to reference T4/T5/T6/T6.5
This includes the renaming of variables, the correction of comments, the correction of print statements from the addon, the correction of table entries in the data from the addon, as well as the UI.

IMPORTANT NOTE: All of these refactors were built on top of the last state of my Classic AddOn, which had re-worked DKP awarding logic from Naxx where only Naxx was awarding DKP and it was doing so for all previous raid tiers. This DKP award logic needs to be refactored to its previous state where each tier awards its own DKP.

Additionally, the AddOn itself houses the logic for determining when certain oddball encounters have been defeated (such as The Prophet Skeram, Razorgore, and The Four Horsemen). This logic needs to be removed, and new logic needs to be added for the High King Maulgar council fight.

Lastly, all loot and bossconfig files still need to be added